[w3] 3주차 Woostagram 기능 구현 (#27)
* fix: fix merge conflict

* refactor: change modal to class

* feature: follow/following

* feature: main page loading when init

* refactor: code format

* refactor: code format

* feature: index progress

* refactor: reduce duplication

* refactor: change url of follow

* refactor: use comment is author

* refactor : config package

* refactor: axios async await remove &

* refactor: feedback

* refactor: change package of base entity

* feat: s3 yaml

* fix: configuration

* fix: axios import

* feat: add application runner

* feat: article update and delete auth check

* refactor: convention edit

* feat: add data in application runner

* feat: service test mod

* feat: fix test cases

* feat: fix test data loading

* feat: fix url policy into hierarchical

* refactor: user controller test

* refactor: makes tests use pre-loaded test data

* feat: add follow test code

* feat : likes service test

* feature: like

* feat: add production data

* fix: harry.jpeg

* refactor: detatch template

* feat: add uploading user profile image

* fix: do not stop service when s3 crash

* fix: profile main view

* refactor: mypage script extract

* feat: add main controller test

* feat: add main service test

* feat : add likes page

* feature: like number

* refactor: remove comments in production code

* feature: head link

* refactor: article test convention mod

# Conflicts:
#	src/main/resources/static/js/comment.js
#	src/main/resources/static/js/common.js

* refactor: article test convention

* refactor: comment test convention

* feat: add follow repo test, convention mod

* feat: add repository test , convention mod

* refactor: user test

* feat: add index test

* refactor: convention mod

* feature: main comment add

* feature add endponint at user controller for getting loggedIn user information

* feat : add show userpage

* feature: get loggedin user information

* feature: comment add realtime

* feature: mypage article rest api

* feat: add STOMP library to gradle

* feat: add websocket library to gradle

* feat : logout

* refactor : followers num

* refactor: update uploading user profile

* feat: add websocket config

* refactor: test error mod

* refactor: add semicolon

* feature: main lazyload

* refactor: update getModalTemplate function in template.js

* refactor: add distinctive number to modal function

* refactor: add uploading profile imgae feature in mypage.html

* refactor: change rest api parameter's dto type to multipartfile

* feat: test code

* feat : user mypage get article

* refactor : user mypage

* feat: image resize

* feat: test code

* feat: likes alarm service

* fix: production data conflict resolved

* feat: self-likes does not issue alarm

* refactor: fix failed test in likes service test

* feat: cannot self-likes

* feat: send follow alarm message

* refactor: set transaction

* refactor: rename mypage html page

* refactor: remove user controller api not used

* refactor: resolve conflict in follow service test

* feat: user search service & controller

* feat: user search service refactor

* feat: hashtag service

* feat: article service saves hash tags

* feat: full search user & hashtag

* fix: article service test add mock

* feat: add user exception handler

* refactor: rename variable in FullSearchService

* refactor: add user content constraints

* fix: user creation constraints

* refactor: code format

* html : user mypage lazy load

* feat: resize retaining image ratio

* refactor: remove duplicated tags

* fix: websocket_cors

* fix: resolve failed tests

* feat: main share

* feat: search frontend

* feat: add user's duplicated constraints

* refactor: change css files

* feat: search frontend

* feat: hash tag backend

* fix: bug resolved

* refactor: configure profile image

* feat: update changing user profiles

* refactor: code format

* refactor: update css

* refactor: update mypage view

* feat: article detail interactions

* feat: aws s3 service test code

* refactor: code format

* feat: alarm service test

* feat: test

* fix: use new aws server address

* refactor: use relative path for websocket connection

* feat: use websocket not only index page

* feat: make exception chained

* refactor: code format

* feat: print stack trace

* fix: please match image type

* fix: please match image type

* refactor: update button mod

* refactor : article search

* refactor : imports optimize

* feat: check duplicated user name when updating

* feat: add url to author name of article

* refactor: change package name

* fix: layout

* fix: markup

* feat: date view

* fix: article edit page load comments

* fix: html render in comment and article

* refactor : console log

* refactor: alarm javascript

* refactor: reflect frontend 3rd review

* fix: index e to element

* fix: prevent log when not localhost

* feat: error handling by alert

* fix: remove bookmark button

* feat: hashtag link

* feat: add exception handling

* refactor: change exception handler's package

* feat: empty image file exception

* fix: error script add

* fix: search uri #

* fix: inform password format

* refactor: style

* feat: rate limiter

* refactor: resolve conflict

* fix: logout error

* refactor: remove else

* refactor: add referer in header when requesting

* fix: test code

* refactor: brown review

* refactor: feedback refelect

* refactor: feedback reflect

* fix: image ratio detail page

* fix: image ratio detail page

* fix: focus out & markup

* feat: makes request rate limiter aop

* refactor: refact code convention

* feat: recommended article

* refactor: refact test code convetion

+++ src/main/resources/static/js/comment.js
@@ -1,4 +1,4 @@
-const Comment = (function () {
+const Comment = (() => {
     const CommentController = function () {
         const commentService = new CommentService();
 
@@ -14,7 +14,7 @@ const Comment = (function () {
             button.addEventListener('click', commentService.remove);
         };
 
-        const init = function () {
+        const init = () => {
             createButton();
             removeButton()
         };
@@ -26,33 +26,31 @@ const Comment = (function () {
     const CommentService = function () {
         const request = new Request(`/api/articles/${articleId}/comments`);
 
-        const commentTemplate =
-            `<div class="contents-inner">
-                <div class="profile">
-                    <img src={{userInfoDto.userContentsDto.profile}}>
+        const getCommentTemplate = (data) =>
+            `<div class="profile">
+                    <img src=${data.userInfoDto.profile}>
                     <div class="profile-text">
-                        <span class="profile-name">{{userInfoDto.userContentsDto.userName}}</span>
-                        <span class="contents-para">{{contents}}</span>
+                        <span class="profile-name"><a href="/${data.userInfoDto.userContentsDto.userName}">
+                            ${data.userInfoDto.userContentsDto.userName}</a></span>
+                        <span class="contents-para">${htmlToStringParse(data.contents)}</span>
                     </div>
-                    <button class="comment-delete" data-id={{id}}>삭제</button>
-                </div>
-            </div>`;
-
-        const commentItemTemplate = Handlebars.compile(commentTemplate);
+                    <div class="comment-delete" data-id=${data.id}>                
+                        <i class=" fa fa-times" aria-hidden="true"></i>
+                    </div>
+                </div>`;
 
         const read = () => {
             request.get('/', (status, data) => {
                 document.querySelector('.comment-list').innerHTML = "";
-
                 data.forEach(e => {
-                    document.querySelector('.comment-list').insertAdjacentHTML('beforeend', commentItemTemplate(e))
+                    document.querySelector('.comment-list').insertAdjacentHTML('beforeend', getCommentTemplate(e))
                 });
             })
         };
 
         const create = () => {
             const commentInput = document.querySelector(".comment-input");
-            let contents = commentInput.value;
+            const contents = commentInput.value;
 
             if (contents.length === 0) {
                 return false;
@@ -67,6 +65,11 @@ const Comment = (function () {
         };
 
         const remove = (event) => {
+            console.log(event.target.classList.contains("comment-delete"));
+            if (!event.target.classList.contains("comment-delete")) {
+                return;
+            }
+
             const commentId = event.target.getAttribute("data-id");
 
             request.delete('/' + commentId, (status, data) => {
@@ -88,6 +91,6 @@ const Comment = (function () {
     return {
         init: init
     }
-}());
+})();
 
 Comment.init();
